Leave Encashment and pay structure fixes

+++ hrms-project/src/app/leave-master/leave.service.ts
@@ -519,6 +519,23 @@ CreateLeaveapprovalLevel(payload: any): Observable<any> {
     );
   }
 
+    CreateLeaveEncashment(formData: FormData): Observable<any> {
+    const selectedSchema = localStorage.getItem('selectedSchema');
+    if (!selectedSchema) {
+      console.error('No schema selected.');
+      return throwError('No schema selected.');
+    }
+  
+    const apiUrl = `${this.apiUrl}/payroll/api/emp-leave-encashment/?schema=${selectedSchema}`;
+  
+    return this.http.post(apiUrl, formData).pipe(
+      catchError((error) => {
+        console.error('Error during leave type registration:', error);
+        return throwError(error);
+      })
+    );
+  }
+
 
 
 
@@ -533,6 +550,21 @@ CreateLeaveapprovalLevel(payload: any): Observable<any> {
   }
  
   const apiUrl = `${this.apiUrl}/calendars/api/Leave_balance/${categoryId}/?schema=${selectedSchema}`;
+ 
+  return this.http.delete(apiUrl);
+}
+
+  deleteLeaveEncashment(categoryId: number): Observable<any> {
+  // const url = `${this.baseUrl}/Catogory/${categoryId}`;
+  // return this.http.delete(url);
+
+  const selectedSchema = localStorage.getItem('selectedSchema');
+  if (!selectedSchema) {
+    console.error('No schema selected.');
+    return throwError('No schema selected.'); // Return an error observable if no schema is selected
+  }
+ 
+  const apiUrl = `${this.apiUrl}/payroll/api/emp-leave-encashment/${categoryId}/?schema=${selectedSchema}`;
  
   return this.http.delete(apiUrl);
 }
@@ -559,6 +591,19 @@ CreateLeaveapprovalLevel(payload: any): Observable<any> {
 updateLeaveBalance(id: number, data: any): Observable<any> {
   const selectedSchema = localStorage.getItem('selectedSchema');
   const apiUrl = `${this.apiUrl}/calendars/api/Leave_balance/${id}/?schema=${selectedSchema}`;
+  const headers = new HttpHeaders({ 'Content-Type': 'application/json' });
+
+  return this.http.put(apiUrl, data, { headers }).pipe(
+    catchError((error) => {
+      console.error('Error updating asset:', error);
+      return throwError(error);
+    })
+  );
+}
+
+updateLeaveEncashment(id: number, data: any): Observable<any> {
+  const selectedSchema = localStorage.getItem('selectedSchema');
+  const apiUrl = `${this.apiUrl}/payroll/api/emp-leave-encashment/${id}/?schema=${selectedSchema}`;
   const headers = new HttpHeaders({ 'Content-Type': 'application/json' });
 
   return this.http.put(apiUrl, data, { headers }).pipe(
@@ -870,6 +915,18 @@ rejectApprovalRequestLeave(apiUrl: string, approvalData: { note: string; status:
     const branchParam = branchIds.length > 0 ? `[${branchIds.join(',')}]` : '';
     
     let url = `${this.apiUrl}/calendars/api/Leave_balance/?schema=${selectedSchema}`;
+    if (branchParam) {
+      url += `&branch_id=${branchParam}`;
+    }
+    
+    return this.http.get(url);
+  }
+
+     getAllLeaveEncashmentAllNew(selectedSchema: string, branchIds: number[]): Observable<any> {
+    // Converts [1,3,4] into the string "[1,3,4]" for the URL
+    const branchParam = branchIds.length > 0 ? `[${branchIds.join(',')}]` : '';
+    
+    let url = `${this.apiUrl}/payroll/api/emp-leave-encashment/?schema=${selectedSchema}`;
     if (branchParam) {
       url += `&branch_id=${branchParam}`;
     }
